+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## UNRELEASED
 - Feature: Enable `producers.persistent_connections` phobos setting
+- Feature: Add consumer configuration, `backfill_associations` to import associated records of primary class prior to upserting primary records
 
 # 1.24.2 - 2024-05-01
 - Fix: Deprecation notice with Rails 7.